Redefine objects locality
local_missing and remote_missing, used to build the next messages to
send, are now local to the stringReceived() method and do not pollute
the whole class.

In addition, an other set (new_hashes) lists all the hashes missing
upon decoding, and removes them upon reception. When the transfer
is over, this set must be empty.